pyproject.toml
[project]
name = "python-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
